Restructure input package
