+++ packages/graphics/src/core/input/InputAction.cpp
@@ -1,4 +1,4 @@
-#include <graphics/input/InputAction.h>
+#include <graphics/core/input/InputAction.h>
 
 InputAction::InputAction(const std::string& name, const InputEventType type) : name(name), type(type)
 {
