+++ packages/graphics/src/core/input/InputController.cpp
@@ -1,4 +1,4 @@
-#include <graphics/core/windows/InputController.h>
+#include <graphics/core/input/InputController.h>
 #include <GLFW/glfw3.h>
 #include <common/Assertions.h>
